lnet: remove WINS_MyAddress / Net_MyAddress
As part of `l_net` there's currently code where a UDP socket is created
just to get the host's IP address, which is then stored and can be
retrieved via the above 2 functions.

`Net_MyAddress` is a shim around `WINS_MyAddress`, and was never used.

`WINS_MyAddress` was only used when `_DEBUG` flag is enabled at build
time to debug-print "the" IP address to the console.

Being that users never saw this message to begin with, it seems safe to
assume today that developers know themselves how to determine "the" IP
Address of their hosts without the need for this being built into
radiant.

This commit removes both functions, along with the UDP socket creation
code.

This also fixes a standing issue that these UDP sockets were never
closed and each build invocation accumulated lingering unconnected UDP
sockets.